rename to useTimeTravel
