+++ src/utils/index.js
@@ -5,4 +5,4 @@ export { default as useShare } from "./useShare";
 export { default as useTitle } from "./useTitle";
 export { default as useWeekDays } from "./useWeekDays";
 export { default as useCalculation } from "./useCalculation";
-export { default as useChartReducer } from "./useChartReducer";
+export { default as useTimeTravel } from "./useTimeTravel";
